wip: introduce Exporter trait
