+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cli;
 pub mod code;
 pub mod constraint;
+pub mod exporter;
 pub mod language;
 pub mod matcher;
 pub mod node;
